:hammer: Changed weight to a double in database and DTOs.

+++ src/main/java/DTOs/Boat.java
@@ -4,7 +4,7 @@ public class Boat extends Vehicle {
     private int numLifeBoats;
     private int max_speed_knots;
 
-    public Boat(int id, String type, String make, String model, String engine, String registration, String color, int weightInTonnes, int numPassengers, int mileage, int price, String fuelType, Dealer dealer, String imgUrl,int numLifeBoats,int max_speed_knots) {
+    public Boat(int id, String type, String make, String model, String engine, String registration, String color, double weightInTonnes, int numPassengers, int mileage, int price, String fuelType, Dealer dealer, String imgUrl,int numLifeBoats,int max_speed_knots) {
         super(id, type, make, model, engine, registration, color, weightInTonnes, numPassengers, mileage, price, fuelType, dealer,imgUrl);
         this.numLifeBoats = numLifeBoats;
         this.max_speed_knots = max_speed_knots;
