+++ src/main/java/DTOs/Car.java
@@ -4,10 +4,11 @@ public class Car extends Vehicle {
 private int numDoors;
 
 
-    public Car(int id, String type,String make, String model, String engine, String registration, String color, int weightInTonnes, int numPassengers, int mileage, int price, String fuelType, Dealer dealer,String imgUrl, int numDoors) {
+    public Car(int id, String type,String make, String model, String engine, String registration, String color, double weightInTonnes, int numPassengers, int mileage, int price, String fuelType, Dealer dealer,String imgUrl, int numDoors) {
         super(id, type, make, model, engine, registration, color, weightInTonnes, numPassengers, mileage, price, fuelType, dealer,imgUrl);
         this.numDoors = numDoors;
     }
+
 
     public int getNumDoors() {
         return numDoors;
