+++ src/main/java/DTOs/Truck.java
@@ -4,7 +4,7 @@ public class Truck extends Vehicle {
 
     private int weight_capacity;
 
-    public Truck(int id, String type, String make, String model, String engine, String registration, String color, int weightInTonnes, int numPassengers, int mileage, int price, String fuelType, Dealer dealer, String imgUrl, int weight_capacity) {
+    public Truck(int id, String type, String make, String model, String engine, String registration, String color, double weightInTonnes, int numPassengers, int mileage, int price, String fuelType, Dealer dealer, String imgUrl, int weight_capacity) {
         super(id, type, make, model, engine, registration, color, weightInTonnes, numPassengers, mileage, price, fuelType, dealer, imgUrl);
         this.weight_capacity = weight_capacity;
     }
